requirements.dev.txt
pytest==3.2.5
pyyaml==3.12
backoff==1.4.3
kubernetes
